Move action_module from module_utils to plugin_utils.
