+++ plugins/action/openssl_privatekey_pipe.py
@@ -11,7 +11,7 @@ import base64
 
 from ansible.module_utils._text import to_native, to_bytes
 
-from ansible_collections.community.crypto.plugins.module_utils.action_module import ActionModuleBase
+from ansible_collections.community.crypto.plugins.plugin_utils.action_module import ActionModuleBase
 
 from ansible_collections.community.crypto.plugins.module_utils.crypto.basic import (
     OpenSSLObjectError,
